Added a few things. See description.
Method to get all Workspaces. Returns a "pretty" string with name and workplace id (WID) .

Function that returns project ID: uses the WID as input and returns a "pretty" string with all project id's PID linked to that workspace.

Split the implementations into two files to keep it more organized.
